Remove relay
